(Bin + Lib + Cli) move commands to own subdir
