+++ bin/flyde.dart
@@ -2,11 +2,11 @@ import 'dart:io';
 
 import 'package:args/command_runner.dart';
 
-import 'package:flyde/features/cli/build_command.dart';
-import 'package:flyde/features/cli/create_command.dart';
-import 'package:flyde/features/cli/init_command.dart';
-import 'package:flyde/features/cli/load_command.dart';
-import 'package:flyde/features/cli/run_command.dart';
+import 'package:flyde/features/cli/command/build_command.dart';
+import 'package:flyde/features/cli/command/create_command.dart';
+import 'package:flyde/features/cli/command/init_command.dart';
+import 'package:flyde/features/cli/command/load_command.dart';
+import 'package:flyde/features/cli/command/run_command.dart';
 
 void main(List<String> args) {
   final runner = CommandRunner('flyde', 'An easy to use C++ distributed compilation framework.')
