+++ lib/features/cli/command/create_command.dart
