+++ lib/features/cli/command/init_command.dart
