+++ lib/features/cli/command/load_binary_command.dart
