+++ lib/features/cli/command/load_command.dart
